Add files via upload
A simple Arduino code to check the water level set by the user with the help of a customizable slider with two heads provided to set the upper and lower thresholds as per the requirements. Also, there are led's incorporated which glow when the water level exceeds or recedes the set threshold value indicating to either turn on or turn off the motor according to the color of the led glowing.

The Html file here shows us a simple representation of a tank and a sump each provided with a sliding handle [customizable as per the user's requirements] placed next to each of the reservoirs.